+++ wxc_sdk/person_settings/msteams.py
@@ -4,9 +4,10 @@ from typing import Optional
 from wxc_sdk.api_child import ApiChild
 from wxc_sdk.base import ApiModel
 
-__all__ = ['MSTeamsSettings', 'SettingsObject', 'MSTeamsSettingApi']
+__all__ = ['MSTeamsSettings', 'SettingsObject', 'MSTeamsSettingApi', 'OrgMSTeamsSettingApi', 'OrgMSTeamsSettings']
 
 
+# noinspection DuplicatedCode
 class SettingsObject(ApiModel):
     #: Name of the setting retrieved.
     setting_name: Optional[str] = None
@@ -22,6 +23,15 @@ class MSTeamsSettings(ApiModel):
     #: Unique identifier for the person.
     person_id: Optional[str] = None
     #: Unique identifier for the organization in which the person resides.
+    org_id: Optional[str] = None
+    #: Array of `SettingsObject`.
+    settings: Optional[list[SettingsObject]] = None
+
+
+class OrgMSTeamsSettings(ApiModel):
+    #: Level at which the `settingName` has been set.
+    level: Optional[str] = None
+    #: Unique identifier for the organization.
     org_id: Optional[str] = None
     #: Array of `SettingsObject`.
     settings: Optional[list[SettingsObject]] = None
@@ -85,4 +95,74 @@ class MSTeamsSettingApi(ApiChild, base='telephony/config/people'):
         body['settingName'] = setting_name
         body['value'] = value
         url = self.ep(f'{person_id}/settings/msTeams')
+        super().put(url, params=params, json=body)
+
+
+class OrgMSTeamsSettingApi(ApiChild, base='telephony/config/settings/msTeams'):
+    """
+    Client Call Settings
+
+    Client Call Settings supports reading and writing of Webex Calling client settings for a specific organization.
+
+    Viewing these read-only organization settings requires a full or read-only administrator auth token with a scope of
+    `spark-admin:telephony_config_read`.
+
+    Modifying these organization settings requires a full administrator auth token with a scope of
+    `spark-admin:telephony_config_write`.
+
+    A partner administrator can retrieve or change settings in a customer's organization using the optional `orgId`
+    query parameter.
+    """
+
+    def read(self, org_id: str = None) -> OrgMSTeamsSettings:
+        """
+        Get an Organization's MS Teams Settings
+
+        Get organization MS Teams settings.
+
+        At an organization level, MS Teams settings allow access to viewing the `HIDE WEBEX APP` and `PRESENCE SYNC`
+        settings.
+
+        To retrieve an organization's MS Teams settings requires a full or read-only administrator auth token with a
+        scope of `spark-admin:telephony_config_read`.
+
+        :param org_id: Retrieve MS Teams settings for the organization.
+        :type org_id: str
+        :rtype: :class:`OrgMSTeamsSettings`
+        """
+        params = {}
+        if org_id is not None:
+            params['orgId'] = org_id
+        url = self.ep()
+        data = super().get(url, params=params)
+        r = OrgMSTeamsSettings.model_validate(data)
+        return r
+
+    def configure(self, setting_name: str, value: bool,
+                  org_id: str = None):
+        """
+        Update an Organization's MS Teams Setting
+
+        Update an MS Teams setting.
+
+        MS Teams setting can be updated at the organization level.
+
+        Requires a full administrator auth token with a scope of `spark-admin:telephony_config_write`.
+
+        :param setting_name: The enum value, either `HIDE_WEBEX_APP` or `PRESENCE_SYNC`, for the respective
+            `settingName` to be updated.
+        :type setting_name: SettingsObjectSettingName
+        :param value: The boolean value, either `true` or `false`, for the respective `settingName` to be updated.
+        :type value: bool
+        :param org_id: Update MS Teams setting value for the organization.
+        :type org_id: str
+        :rtype: None
+        """
+        params = {}
+        if org_id is not None:
+            params['orgId'] = org_id
+        body = dict()
+        body['settingName'] = setting_name
+        body['value'] = value
+        url = self.ep()
         super().put(url, params=params, json=body)